Moved compose views to their own file
