+++ WordPress/src/main/java/org/wordpress/android/ui/main/MeEmailVerificationBannerCompose.kt
@@ -1,12 +1,6 @@
 package org.wordpress.android.ui.main
 
-import android.animation.Animator
-import android.animation.AnimatorListenerAdapter
-import android.animation.ObjectAnimator
-import android.content.Context
-import android.util.AttributeSet
-import android.view.animation.LinearInterpolator
-import android.widget.FrameLayout
+import android.content.res.Configuration
 import androidx.compose.foundation.clickable
 import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.Row
@@ -20,143 +14,17 @@ import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
-import androidx.compose.ui.platform.ComposeView
 import androidx.compose.ui.res.colorResource
 import androidx.compose.ui.res.painterResource
 import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.text.font.FontWeight
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
-import androidx.core.view.isVisible
 import org.wordpress.android.R
 
 /**
- * Custom view for Me screen email verification banner
+ * TODO Dark mode support
  */
-class MeEmailVerificationBanner @JvmOverloads constructor(
-    context: Context,
-    attrs: AttributeSet? = null,
-    defStyleAttr: Int = 0
-) : FrameLayout(context, attrs, defStyleAttr) {
-    private val composeView: ComposeView = ComposeView(context)
-
-    private var verificationState: MeViewModel.EmailVerificationState? = null
-    private var emailAddress: String = ""
-    private var onLinkClick: (context: Context) -> Unit = {}
-
-    private val animDuration =
-        context.resources.getInteger(android.R.integer.config_shortAnimTime).toLong()
-
-    init {
-        addView(composeView)
-    }
-
-    fun setVerificationState(
-        verificationState: MeViewModel.EmailVerificationState,
-        emailAddress: String = "",
-        onLinkClick: (context: Context) -> Unit = {},
-    ) {
-        this.verificationState = verificationState
-        this.emailAddress = emailAddress
-        this.onLinkClick = onLinkClick
-
-        when (verificationState) {
-            MeViewModel.EmailVerificationState.UNVERIFIED,
-            MeViewModel.EmailVerificationState.LINK_REQUESTED,
-            MeViewModel.EmailVerificationState.LINK_SENT,
-            MeViewModel.EmailVerificationState.LINK_ERROR -> {
-                if (isVisible) {
-                    fadeOut(true)
-                } else {
-                    fadeInAndUpdate()
-                }
-            }
-
-            MeViewModel.EmailVerificationState.VERIFIED,
-            MeViewModel.EmailVerificationState.NO_ACCOUNT -> {
-                if (isVisible.not()) {
-                    fadeOut(false)
-                }
-            }
-        }
-    }
-
-    private fun updateContent() {
-        when (verificationState!!) {
-            MeViewModel.EmailVerificationState.UNVERIFIED -> {
-                composeView.setContent {
-                    MeEmailUnverifiedBanner(
-                        emailAddress = emailAddress,
-                        onSendLinkClick = {
-                            onLinkClick(context)
-                        }
-                    )
-                }
-            }
-
-            MeViewModel.EmailVerificationState.LINK_REQUESTED -> {
-                composeView.setContent {
-                    MeEmailVerificationSendingBanner(
-                        emailAddress = emailAddress
-                    )
-                }
-            }
-
-            MeViewModel.EmailVerificationState.LINK_SENT -> {
-                composeView.setContent {
-                    MeEmailVerificationSentBanner(
-                        emailAddress = emailAddress
-                    )
-                }
-            }
-
-            MeViewModel.EmailVerificationState.LINK_ERROR -> {
-                composeView.setContent {
-                    MeEmailVerificationErrorBanner(
-                        onResendLinkClick = {
-                            onLinkClick(context)
-                        }
-                    )
-                }
-            }
-
-            else -> {
-                // do nothing
-            }
-        }
-    }
-
-    private fun fadeInAndUpdate() {
-        with(ObjectAnimator.ofFloat(this, ALPHA, 0.0f, 1.0f)) {
-            setDuration(animDuration)
-            interpolator = LinearInterpolator()
-            addListener(object : AnimatorListenerAdapter() {
-                override fun onAnimationStart(animation: Animator) {
-                    this@MeEmailVerificationBanner.isVisible = true
-                    updateContent()
-                }
-            })
-            start()
-        }
-    }
-
-    private fun fadeOut(updateAfterFadeOut: Boolean) {
-        with(ObjectAnimator.ofFloat(this, ALPHA, 1.0f, 0.0f)) {
-            setDuration(animDuration)
-            interpolator = LinearInterpolator()
-            addListener(object : AnimatorListenerAdapter() {
-                override fun onAnimationEnd(animation: Animator) {
-                    if (updateAfterFadeOut) {
-                        fadeInAndUpdate()
-                    } else {
-                        this@MeEmailVerificationBanner.isVisible = false
-                    }
-                }
-            })
-            start()
-        }
-    }
-}
 
 /**
  * Banner when user's email hasn't yet been verified
@@ -231,7 +99,7 @@ fun MeEmailVerificationSendingBanner(
             )
         }
 
-        Row() {
+        Row {
             Text(
                 text = stringResource(
                     R.string.me_email_verification_sending_description,
@@ -275,6 +143,7 @@ fun MeEmailVerificationSentBanner(
                 emailAddress
             ),
             style = MaterialTheme.typography.bodyLarge,
+            color = MaterialTheme.colorScheme.onSurface,
             modifier = Modifier.padding(top = 8.dp)
         )
     }
@@ -337,6 +206,7 @@ private fun MeEmailVerificationContainer(
 }
 
 @Preview
+@Preview(uiMode = Configuration.UI_MODE_NIGHT_YES)
 @Composable
 fun MeEmailUnverifiedPreview() {
     MeEmailUnverifiedBanner(
@@ -346,6 +216,7 @@ fun MeEmailUnverifiedPreview() {
 }
 
 @Preview
+@Preview(uiMode = Configuration.UI_MODE_NIGHT_YES)
 @Composable
 fun MeEmailVerificationRequestedPreview() {
     MeEmailVerificationSendingBanner(
@@ -354,6 +225,7 @@ fun MeEmailVerificationRequestedPreview() {
 }
 
 @Preview
+@Preview(uiMode = Configuration.UI_MODE_NIGHT_YES)
 @Composable
 fun MeEmailVerificationSentPreview() {
     MeEmailVerificationSentBanner(
@@ -362,6 +234,7 @@ fun MeEmailVerificationSentPreview() {
 }
 
 @Preview
+@Preview(uiMode = Configuration.UI_MODE_NIGHT_YES)
 @Composable
 fun MeEmailVerificationErrorPreview() {
     MeEmailVerificationErrorBanner(
